+++ plugins/vitepress-plugin-catalogue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # vitepress-plugin-catalogue
 
+## 1.0.8
+
+### Patch Changes
+
+- process.cwd() 替换为 srcDir
+
 ## 1.0.7
 
 ### Patch Changes